requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
motor>=3.1
pydantic[email]>=1.10
python-jose>=3.3.0
passlib[bcrypt]>=1.7
bcrypt==3.2.0
python-dotenv>=1.0
